requirements.txt
Django==3.2.20
tensorflow==2.15.0
keras==2.15.0
numpy==1.24.3
scikit-learn==1.3.0
opencv-python==4.8.0.76
pillow==9.5.0
matplotlib==3.7.1
django-crispy-forms==2.0
django-bootstrap4==22.2
seaborn==0.12.2
jupyter==1.0.0
